GH-2925 renamed elasticsearch-store integration tests to new pattern
