+++ core/sail/elasticsearch-store/src/test/java/org/eclipse/rdf4j/sail/elasticsearchstore/compliance/ElasticsearchStoreContextIT.java
@@ -13,8 +13,7 @@ import java.io.IOException;
 import org.assertj.core.util.Files;
 import org.eclipse.rdf4j.sail.NotifyingSail;
 import org.eclipse.rdf4j.sail.NotifyingSailConnection;
-import org.eclipse.rdf4j.sail.SailConcurrencyTest;
-import org.eclipse.rdf4j.sail.SailException;
+import org.eclipse.rdf4j.sail.RDFNotifyingStoreTest;
 import org.eclipse.rdf4j.sail.elasticsearchstore.ElasticsearchStore;
 import org.eclipse.rdf4j.sail.elasticsearchstore.SingletonClientProvider;
 import org.eclipse.rdf4j.sail.elasticsearchstore.TestHelpers;
@@ -24,14 +23,15 @@ import org.junit.BeforeClass;
 import pl.allegro.tech.embeddedelasticsearch.EmbeddedElastic;
 
 /**
- * An extension of {@link SailConcurrencyTest} for testing the class
+ * An extension of RDFStoreTest for testing the class
  * {@link org.eclipse.rdf4j.sail.elasticsearchstore.ElasticsearchStore}.
  */
-public class ElasticsearchStoreConcurrencyTest extends SailConcurrencyTest {
+public class ElasticsearchStoreContextIT extends RDFNotifyingStoreTest {
 
 	private static EmbeddedElastic embeddedElastic;
 
 	private static File installLocation = Files.newTemporaryFolder();
+
 	private static SingletonClientProvider clientPool;
 
 	@BeforeClass
@@ -50,18 +50,14 @@ public class ElasticsearchStoreConcurrencyTest extends SailConcurrencyTest {
 
 	}
 
-	/*---------*
-	 * Methods *
-	 *---------*/
-
 	@Override
-	protected NotifyingSail createSail() throws SailException {
-		NotifyingSail sail = new ElasticsearchStore(clientPool, "index1");
-		try (NotifyingSailConnection connection = sail.getConnection()) {
+	protected NotifyingSail createSail() {
+		ElasticsearchStore elasticsearchStore = new ElasticsearchStore(clientPool, "index1");
+		try (NotifyingSailConnection connection = elasticsearchStore.getConnection()) {
 			connection.begin();
 			connection.clear();
 			connection.commit();
 		}
-		return sail;
+		return elasticsearchStore;
 	}
 }
