+++ core/sail/elasticsearch-store/src/test/java/org/eclipse/rdf4j/sail/elasticsearchstore/compliance/ElasticsearchStoreInterruptIT.java
@@ -11,9 +11,10 @@ import java.io.File;
 import java.io.IOException;
 
 import org.assertj.core.util.Files;
-import org.eclipse.rdf4j.repository.Repository;
-import org.eclipse.rdf4j.repository.SparqlOrderByTest;
-import org.eclipse.rdf4j.repository.sail.SailRepository;
+import org.eclipse.rdf4j.sail.NotifyingSail;
+import org.eclipse.rdf4j.sail.NotifyingSailConnection;
+import org.eclipse.rdf4j.sail.SailConcurrencyTest;
+import org.eclipse.rdf4j.sail.SailInterruptTest;
 import org.eclipse.rdf4j.sail.elasticsearchstore.ElasticsearchStore;
 import org.eclipse.rdf4j.sail.elasticsearchstore.SingletonClientProvider;
 import org.eclipse.rdf4j.sail.elasticsearchstore.TestHelpers;
@@ -22,11 +23,16 @@ import org.junit.BeforeClass;
 
 import pl.allegro.tech.embeddedelasticsearch.EmbeddedElastic;
 
-public class ElasticsearchStoreSparqlOrderByTest extends SparqlOrderByTest {
+/**
+ * An extension of {@link SailConcurrencyTest} for testing the class
+ * {@link org.eclipse.rdf4j.sail.elasticsearchstore.ElasticsearchStore}.
+ */
+public class ElasticsearchStoreInterruptIT extends SailInterruptTest {
 
 	private static EmbeddedElastic embeddedElastic;
 
 	private static File installLocation = Files.newTemporaryFolder();
+
 	private static SingletonClientProvider clientPool;
 
 	@BeforeClass
@@ -46,10 +52,13 @@ public class ElasticsearchStoreSparqlOrderByTest extends SparqlOrderByTest {
 	}
 
 	@Override
-	protected Repository newRepository() throws IOException {
-		SailRepository sailRepository = new SailRepository(
-				new ElasticsearchStore(clientPool, "index1"));
-		return sailRepository;
+	protected NotifyingSail createSail() {
+		ElasticsearchStore elasticsearchStore = new ElasticsearchStore(clientPool, "index1");
+		try (NotifyingSailConnection connection = elasticsearchStore.getConnection()) {
+			connection.begin();
+			connection.clear();
+			connection.commit();
+		}
+		return elasticsearchStore;
 	}
-
 }
