+++ core/sail/elasticsearch-store/src/test/java/org/eclipse/rdf4j/sail/elasticsearchstore/compliance/ElasticsearchStoreSparqlRegexIT.java
@@ -8,13 +8,12 @@
 package org.eclipse.rdf4j.sail.elasticsearchstore.compliance;
 
 import java.io.File;
+import java.io.IOException;
 
 import org.assertj.core.util.Files;
-import org.eclipse.rdf4j.sail.NotifyingSail;
-import org.eclipse.rdf4j.sail.NotifyingSailConnection;
-import org.eclipse.rdf4j.sail.Sail;
-import org.eclipse.rdf4j.sail.SailException;
-import org.eclipse.rdf4j.sail.SailIsolationLevelTest;
+import org.eclipse.rdf4j.repository.Repository;
+import org.eclipse.rdf4j.repository.SparqlRegexTest;
+import org.eclipse.rdf4j.repository.sail.SailRepository;
 import org.eclipse.rdf4j.sail.elasticsearchstore.ElasticsearchStore;
 import org.eclipse.rdf4j.sail.elasticsearchstore.SingletonClientProvider;
 import org.eclipse.rdf4j.sail.elasticsearchstore.TestHelpers;
@@ -23,22 +22,16 @@ import org.junit.BeforeClass;
 
 import pl.allegro.tech.embeddedelasticsearch.EmbeddedElastic;
 
-/**
- * An extension of {@link SailIsolationLevelTest} for testing the class
- * {@link org.eclipse.rdf4j.sail.elasticsearchstore.ElasticsearchStore}.
- */
-public class ElasticsearchStoreIsolationLevelTest extends SailIsolationLevelTest {
+public class ElasticsearchStoreSparqlRegexIT extends SparqlRegexTest {
 
 	private static EmbeddedElastic embeddedElastic;
 
 	private static File installLocation = Files.newTemporaryFolder();
-
 	private static SingletonClientProvider clientPool;
 
 	@BeforeClass
-	public static void beforeClass() throws Exception {
+	public static void beforeClass() throws IOException, InterruptedException {
 
-		SailIsolationLevelTest.setUpClass();
 		embeddedElastic = TestHelpers.startElasticsearch(installLocation);
 		clientPool = new SingletonClientProvider("localhost", embeddedElastic.getTransportTcpPort(), "cluster1");
 
@@ -53,14 +46,9 @@ public class ElasticsearchStoreIsolationLevelTest extends SailIsolationLevelTest
 	}
 
 	@Override
-	protected Sail createSail() throws SailException {
-		NotifyingSail sail = new ElasticsearchStore(clientPool, "index1");
-		try (NotifyingSailConnection connection = sail.getConnection()) {
-			connection.begin();
-			connection.clear();
-			connection.commit();
-		}
-		return sail;
+	protected Repository newRepository() throws IOException {
+		SailRepository sailRepository = new SailRepository(
+				new ElasticsearchStore(clientPool, "index1"));
+		return sailRepository;
 	}
-
 }
