+++ core/sail/elasticsearch-store/src/test/java/org/eclipse/rdf4j/sail/elasticsearchstore/compliance/ElasticsearchStoreTupleQueryResultIT.java
@@ -11,9 +11,9 @@ import java.io.File;
 import java.io.IOException;
 
 import org.assertj.core.util.Files;
-import org.eclipse.rdf4j.sail.NotifyingSail;
-import org.eclipse.rdf4j.sail.NotifyingSailConnection;
-import org.eclipse.rdf4j.sail.RDFNotifyingStoreTest;
+import org.eclipse.rdf4j.repository.Repository;
+import org.eclipse.rdf4j.repository.TupleQueryResultTest;
+import org.eclipse.rdf4j.repository.sail.SailRepository;
 import org.eclipse.rdf4j.sail.elasticsearchstore.ElasticsearchStore;
 import org.eclipse.rdf4j.sail.elasticsearchstore.SingletonClientProvider;
 import org.eclipse.rdf4j.sail.elasticsearchstore.TestHelpers;
@@ -22,16 +22,11 @@ import org.junit.BeforeClass;
 
 import pl.allegro.tech.embeddedelasticsearch.EmbeddedElastic;
 
-/**
- * An extension of RDFStoreTest for testing the class
- * {@link org.eclipse.rdf4j.sail.elasticsearchstore.ElasticsearchStore}.
- */
-public class ElasticsearchStoreContextTest extends RDFNotifyingStoreTest {
+public class ElasticsearchStoreTupleQueryResultIT extends TupleQueryResultTest {
 
 	private static EmbeddedElastic embeddedElastic;
 
 	private static File installLocation = Files.newTemporaryFolder();
-
 	private static SingletonClientProvider clientPool;
 
 	@BeforeClass
@@ -51,13 +46,8 @@ public class ElasticsearchStoreContextTest extends RDFNotifyingStoreTest {
 	}
 
 	@Override
-	protected NotifyingSail createSail() {
-		ElasticsearchStore elasticsearchStore = new ElasticsearchStore(clientPool, "index1");
-		try (NotifyingSailConnection connection = elasticsearchStore.getConnection()) {
-			connection.begin();
-			connection.clear();
-			connection.commit();
-		}
-		return elasticsearchStore;
+	protected Repository newRepository() throws IOException {
+		SailRepository sailRepository = new SailRepository(new ElasticsearchStore(clientPool, "index1"));
+		return sailRepository;
 	}
 }
